+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+* Added `global.sfreleaser-min-version` configuration value to force users to upgrade to a new version of `sfreleaser`.
+
 ## v0.6.0
 
 * Added `sfreleaser build` to build artifacts, `sfreleaser build --help` for all the juicy details of the new command.